Add test for existing configuration

+++ commands/svc.cmd
@@ -75,9 +75,14 @@ if [[ "${WARDEN_PARAMS[0]}" == "up" ]]; then
         "$WARDEN_BIN" sign-certificate "${WARDEN_SERVICE_DOMAIN}"
     fi
 
+    if [[ ! -d "${WARDEN_HOME_DIR}/etc/traefik" ]]; then
+        mkdir -p "${WARDEN_HOME_DIR}/etc/traefik"
+    fi
+
     ## copy configuration files into location where they'll be mounted into containers from
-    mkdir -p "${WARDEN_HOME_DIR}/etc/traefik"
-    cp "${WARDEN_DIR}/config/traefik/traefik.yml" "${WARDEN_HOME_DIR}/etc/traefik/traefik.yml"
+    if [[ ! -f "${WARDEN_HOME_DIR}/etc/traefik/traefik.yml" ]]; then
+        cp "${WARDEN_DIR}/config/traefik/traefik.yml" "${WARDEN_HOME_DIR}/etc/traefik/traefik.yml"
+    fi
 
     ## generate dynamic traefik ssl termination configuration
     cat > "${WARDEN_HOME_DIR}/etc/traefik/dynamic.yml" <<-EOT
